refactor: 삭제 여부를 물어보는 로직 추가

+++ section12/src/pages/Edit.jsx
@@ -16,7 +16,10 @@ const Edit = () => {
   };
 
   const onClickDeleteButton = () => {
-    onDelete(params.id);
+    if (window.confirm('정말로 삭제하시겠습니까? 다시 복구되지 않습니다.')) {
+      onDelete(params.id);
+      navigate('/', { replace: true });
+    }
   };
 
   const onSubmit = () => {
